+++ git-remote-updater.php
@@ -11,17 +11,17 @@
 /**
  * Plugin Name:       Git Remote Updater
  * Plugin URI:        https://github.com/afragen/git-remote-updater
- * Description:       Allows you to easily update GitHub Updater repositories in bulk via RESTful endpoint updating.
+ * Description:       Allows you to easily update GitHub Updater repositories in bulk via REST API endpoint updating.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           0.3.4
+ * Version:           0.3.6
  * License:           MIT
  * Network:           true
  * Domain Path:       /languages
  * Text Domain:       git-remote-updater
  * GitHub Plugin URI: https://github.com/afragen/git-remote-updater
  * Requires PHP:      7.1
- * Requires WP:       5.1
+ * Requires at least: 5.2
  */
 
 namespace Fragen\Git_Remote_Updater;
@@ -34,6 +34,7 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-// Setup plugin loading.
-require_once __DIR__ . '/src/Git_Remote_Updater/Bootstrap.php';
+// Load Autoloader.
+require_once __DIR__ . '/vendor/autoload.php';
+
 ( new Bootstrap( __FILE__ ) )->run();