+++ CHANGES.md
@@ -10,3 +10,4 @@
 * more error checking during update process
 * rebrand to Git Remote Updater
 * keep retrying if remote response is WP_Error
+* added filter and function to remove slugs from appearing for updates